refactor: extract thread-agnostic scanning logic into scanCore
